BOTS/03.Website_URL_Summarization/requirements.txt
dotenv
streamlit
langchain
langchain_community
langchain-google-genai
unstructured
google-generativeai
